+++ contrib/opencensus-ext-requests/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
 ## Unreleased
+- Added attributes following specs listed [here](https://github.com/census-instrumentation/opencensus-specs/blob/master/trace/HTTP.md#attributes)
+  ([#746](https://github.com/census-instrumentation/opencensus-python/pull/746))
+- Fixed span name
+  ([#746](https://github.com/census-instrumentation/opencensus-python/pull/746))
 
 ## 0.7.1
 Released 2019-08-06